Fleshed out a rough IAuthorizationService implementation

+++ PackageRepository/Models/ObjectIdentifier.cs
@@ -2,11 +2,11 @@
 
 namespace PackageRepository.Models {
     public class ObjectIdentifier {
-        public int OrganisationId { get; }
+        public long OrganisationId { get; }
         public ObjectType Type { get; }
         public string Name { get; }
 
-        public ObjectIdentifier(int organistionId, ObjectType type, string name) {
+        public ObjectIdentifier(long organistionId, ObjectType type, string name) {
             OrganisationId = organistionId;
             Type = type;
             Name = name;
@@ -14,6 +14,7 @@ namespace PackageRepository.Models {
 
         public override int GetHashCode() {
             var hash = 17;
+
             hash = hash * 23 + OrganisationId.GetHashCode();
             hash = hash * 23 + Type.GetHashCode();
             hash = hash * 23 + Name.GetHashCode();
@@ -22,7 +23,10 @@ namespace PackageRepository.Models {
         }
 
         public override bool Equals(object obj) {
-            return obj is ObjectIdentifier id && id.OrganisationId == OrganisationId && id.Type == Type && id.Name == Name;
+            return obj is ObjectIdentifier id
+                && id.OrganisationId == OrganisationId
+                && id.Type == Type
+                && id.Name == Name;
         }
 
         public static bool operator ==(ObjectIdentifier id1, ObjectIdentifier id2) {
